ai(demo)/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-generativeai
pydantic
